Prep for separate consultation feedback & outcome
This is the first step towards adding support for adding public
feedback to consultations. This commit makes Response the base class
for ConsultationOutcome and updates the UI/controller code so that
we are now dealing with outcomes, not responses. The next step will
be to introduce the ConsultationFeedback model, which will be managed
in much the same way as outcomes.