Introduce tabbed management for document series
Document series documents are now directly managed through the
document series, and are no longer added/removed from the documents
themselves. A search interface allows for documents to be found and
added and removed from the series. It was intended that this would
be an AJAX-powered autocompleting search, but the response time for
searches is just not fast enough without a full-text search engine
behind the scenes. Currently, there is no fallback for non-JS browsers