Change rummager seach backend to not hit the database
This improves the performance as we do not need to load all the editions
and their translations from the database.

We have to still fetch some models from the database such as organisations,
document_series, and operation fields, but this can be refactored later to
add extra data to the seach index.

We see requests times drop from 3-4 seconds to under one second.